urban-land-backend/requirements.txt
asgiref==3.9.1
channels==4.3.2
channels_redis==4.3.0
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.2
mysqlclient==2.2.7
psycopg2-binary==2.9.11
PyJWT==2.10.1
redis==7.1.0
sqlparse==0.5.3
tzdata==2025.2
